api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.339
openai==1.3.5
python-dotenv==1.0.0
cassio==0.1.3
astrapy==0.7.4
pydantic==2.5.2
python-multipart==0.0.6
